Split out code related to pr comments
